+++ random.cpp
@@ -79,9 +79,9 @@ struct SolveToTimelimit : fm::solve_to::Callback
 fm::Matrix random_elimination(fm::System system, int num_drop)
 {
     fm::Matrix& matrix = system.ineqs;
-    size_t num_vars = matrix[0].size();
-    size_t width = intlog2(num_vars)/2;
-    size_t solve_to = 1<<width;
+    int num_vars = matrix[0].size();
+    int width = intlog2(num_vars)/2;
+    int solve_to = 1<<width;
 
     std::random_device rd;
     std::default_random_engine random_engine(rd());
@@ -297,7 +297,7 @@ try
     }
     util::AutogenNotice gen(argc, argv);
 
-    size_t num_drop = atol(argv[1]);;
+    int num_drop = atol(argv[1]);;
     fm::System init_state = fm::parse_matrix(util::read_file(argv[2]));
     fm::System ref_solution = fm::parse_matrix(util::read_file(argv[3]));
     int num_turns = 100;
